completed customer profile endpoint. refactored entities - no longer keyless.

+++ CompanySalesAPI/CompanySalesAPI/Models/DTOs/SaleDto.cs
@@ -3,8 +3,8 @@
     public class SaleDto
     {
         public required string OrderNumber { get; set; }
-        public int ProductKey { get; set; }
-        public long CustomerId { get; set; }
+        public long ProductKey { get; set; }
+        public int CustomerId { get; set; }
         public DateTime OrderDate { get; set; }
         public DateTime ShippingDate { get; set; }
         public DateTime DueDate { get; set; }
